Add EndpointTrait
