+++ src/Illuminate/Routing/Annotations/ResourceEndpoint.php
@@ -4,6 +4,8 @@ use Illuminate\Support\Collection;
 
 class ResourceEndpoint implements EndpointInterface {
 
+	use EndpointTrait;
+
 	/**
 	 * All of the resource controller methods.
 	 *
@@ -177,27 +179,6 @@ class ResourceEndpoint implements EndpointInterface {
 	}
 
 	/**
-	 * Determine if the middleware applies to a given method.
-	 *
-	 * @param  string  $method
-	 * @param  array  $middleware
-	 * @return bool
-	 */
-	protected function middlewareAppliesToMethod($method, array $middleware)
-	{
-		if ( ! empty($middleware['only']) && ! in_array($method, $middleware['only']))
-		{
-			return false;
-		}
-		elseif ( ! empty($middleware['except']) && in_array($method, $middleware['except']))
-		{
-			return false;
-		}
-
-		return true;
-	}
-
-	/**
 	 * Get the names for the given path.
 	 *
 	 * @param  ResourcePath  $path
@@ -219,28 +200,6 @@ class ResourceEndpoint implements EndpointInterface {
 	}
 
 	/**
-	 * Get the controller method for the given endpoint path.
-	 *
-	 * @param  AbstractPath  $path
-	 * @return string
-	 */
-	public function getMethodForPath(AbstractPath $path)
-	{
-		return $path->method;
-	}
-
-	/**
-	 * Add the given path definition to the endpoint.
-	 *
-	 * @param  AbstractPath  $path
-	 * @return void
-	 */
-	public function addPath(AbstractPath $path)
-	{
-		$this->paths[] = $path;
-	}
-
-	/**
 	 * Get all of the path definitions for an endpoint.
 	 *
 	 * @return array[AbstractPath]
@@ -248,31 +207,6 @@ class ResourceEndpoint implements EndpointInterface {
 	public function getPaths()
 	{
 		return $this->paths;
-	}
-
-	/**
-	 * Implode the given list into a comma separated string.
-	 *
-	 * @param  array  $list
-	 * @return string
-	 */
-	protected function implodeArray(array $array)
-	{
-		$results = [];
-
-		foreach ($array as $key => $value)
-		{
-			if (is_string($key))
-			{
-				$results[] = "'".$key."' => '".$value."'";
-			}
-			else
-			{
-				$results[] = "'".$value."'";
-			}
-		}
-
-		return count($results) > 0 ? implode(', ', $results) : '';
 	}
 
 	/**
